DRTII-974 Refactor slick db code & split into class files
