--- server/src/main/scala/slickdb/VoyageManifestPassengerInfoTable.scala
@@ -1,14 +1,0 @@
-package slickdb
-
-import org.slf4j.{Logger, LoggerFactory}
-
-
-case class VoyageManifestPassengerInfoTable(tables: Tables) {
-  val log: Logger = LoggerFactory.getLogger(getClass)
-
-  import tables.profile.api._
-  import tables.VoyageManifestPassengerInfo
-
-  val db: tables.profile.backend.DatabaseDef = Database.forConfig("aggregated-db")
-  val voyageManifestPassengerInfoTableQuery = TableQuery[VoyageManifestPassengerInfo]
-}
