+++ server/src/test/resources/application.conf
@@ -21,7 +21,7 @@ akka {
   }
 }
 
-aggregated-db {
+h2-aggregated-db {
   connectionPool = disabled
   driver = "org.h2.Driver"
   url = "jdbc:h2:mem:drt;DB_CLOSE_DELAY=-1"
