+++ server/src/test/scala/drt/server/feeds/api/DbManifestProcessorTest.scala
@@ -1,0 +1,7 @@
+package drt.server.feeds.api
+
+import org.specs2.mutable.Specification
+
+class DbManifestProcessorTest extends Specification {
+
+}
